Create safe display component for folder list
